backend/requirements.txt
asgiref
dj-database-url
Django
djangorestframework
djangorestframework_simplejwt
psycopg2-binary
PyJWT
python-dotenv
sqlparse
typing_extensions
tzdata
python-decouple
django-cors-headers
whitenoise
openai